pyproject.toml
[project]
name = "telegram-accounting-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.18.0",
    "aiomysql>=0.2.0",
    "aioredis>=2.0.1",
    "fastapi>=0.115.11",
    "jinja2>=3.1.6",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]
